requirements.txt
amqp==2.5.1
appnope==0.1.0
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
decorator==4.4.0
Django==2.2.4
django-extensions==2.2.1
django-preferences==1.0.0
et-xmlfile==1.0.1
importlib-metadata==0.19
ipython==7.7.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
kombu==4.6.4
openpyxl==2.6.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
redis==3.3.8
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
vine==1.3.0
wcwidth==0.1.7
zipp==0.5.2
